% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addComment.R
\name{addComment}
\alias{addComment}
\title{Adds a comment to a thread on Labstep}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` commande}

\item{object}{The labstep object you want to comment on. Can be `protocol`, `experiment` or `resource`}

\item{body}{The body of the comment in plain text.}
}
\value{
Returns a `comment` object
}
\description{
This function allows you to comment on different content types
}
\examples{
user <- login("demo@labstep.com","demopassword")
expName = 'My First Experiment'
expDesc = 'An experiment testing the labstepR library'
experiment <- createExperiment(user,expName,expDesc)
commentBody = 'An experiment testing the labstepR library'
comment <- addComment(user,experiment,commentBody)
print(comment)
}
