% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTagTo.R
\name{addTagTo}
\alias{addTagTo}
\title{Attach an existing tag to an entity on Labstep. See 'tag' for a simplified tagging function.}
\usage{
addTagTo(user, entity, tag)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{entity}{The Labstep entity to tag. Can be a Resource, Experiment or Protocol.}

\item{tag}{A tag entity. Must have an 'id' field.}
}
\value{
Returns the tagged entity.
}
\description{
This function allows you to attach a tag to a Labstep entity (Protocol, Resource or Experiment).
}
\examples{
user <- login("demo@labstep.com","demopassword")
tags <- getTags(user,count=1)
tag = tags[[1]]
experiments <- getExperiments(user,count=1)
experiment = experiments[[1]]
experiment = addTagTo(user,experiment,tag)
}
