% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachFile.R
\name{attachFile}
\alias{attachFile}
\title{Attaches a file to a thread on Labstep}
\usage{
attachFile(user, object, filepath, caption)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api\_key` field. Returned from `login` command}

\item{object}{The labstep object you want to comment on. Can be `protocol`, `experiment` or `resource`}

\item{filepath}{The path to ther file to attach.}

\item{caption}{A caption describing your file (optional)}
}
\value{
Returns a `comment` object with the file attached
}
\description{
This function allows you to attach a file to different content types
}
\examples{
user <- login("demo@labstep.com","demopassword")
expName = 'My First Experiment'
expDesc = 'An experiment testing the labstepR library'
experiment <- createExperiment(user,expName,expDesc)
filepath = 'mydata.RData'
save.image(filepath)
caption = 'The R workspace after running this script'
file <- attachFile(user,experiment,filepath,caption)
print(file)
}
