% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachProtocol.R
\name{attachProtocol}
\alias{attachProtocol}
\title{Attach a Protocol to an Experiment on Labstep}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `authenticate` command}

\item{experiment}{The experiment you want to attach the protocol to. Must have at least an `id` field.}

\item{protocol}{The protocol you want to attach. Must have at least an `id` field}
}
\value{
Returns an `experiment` object
}
\description{
This function allows you to attach a Protocol to an existing Experiment.
}
\examples{
user <- authenticate("demo@labstep.com","demopassword")
experiment <- createExperiment(user,'My First Experiment','An experiment testing the labstepR library')
protocol <- createProtocol(user,'My First Protocol')
experiment <- attachProtocol(user,experiment,protocol)
print(experiment)
}
