% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExperiment.R
\name{createExperiment}
\alias{createExperiment}
\title{Create an Experiment on Labstep}
\usage{
createExperiment(user, name, description = NULL, protocolIds = NULL)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api\_key` field. Returned from `login` command}

\item{name}{The name you want to give your experiment.}

\item{description}{Give your experiment a description to help you remember what it was about.}

\item{protocolIds}{List of ids for protocols to attach to the experiment.}
}
\value{
Returns an `experiment` object
}
\description{
This function allows you to create an experiment card.
}
\examples{
user <- login("demo@labstep.com","demopassword")
expName = 'My First Experiment'
expDesc = 'An experiment testing the labstepR library'
experiment <- createExperiment(user,expName,expDesc)
print(experiment)
}
