% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{Create a Project on Labstep}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{name}{The name you want to give your Project.}
}
\value{
Returns an `project` object
}
\description{
This function allows you to create a Labstep Project.
}
\examples{
user <- login("demo@labstep.com","demopassword")
project <- createProject(user,'My First Project')
print(project)
}
