% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProtocol.R
\name{createProtocol}
\alias{createProtocol}
\title{Create a Protocol on Labstep}
\usage{
createProtocol(user, name)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api\_key` field. Returned from `login` command}

\item{name}{The name you want to give your protocol.}
}
\value{
Returns an `protocol` object
}
\description{
This function allows you to create a protocol.
}
\examples{
user <- login("demo@labstep.com","demopassword")
protocol <- createProtocol(user,'My First Protocol')
print(protocol)
}
