% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResource.R
\name{createResource}
\alias{createResource}
\title{Create a Resource on Labstep}
\usage{
createResource(user, name)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command.}

\item{name}{The name you want to give your resource}
}
\value{
Returns a `resource` object
}
\description{
This function allows you to create a resource on Labstep.
}
\examples{
user <- login("demo@labstep.com","demopassword")
resource <- createResource(user,'My First Resource')
print(resource)
}
