% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTag.R
\name{createTag}
\alias{createTag}
\title{Create a new tag on Labstep}
\usage{
createTag(user, name)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{name}{The name of the tag.}
}
\value{
Returns an `tag` object
}
\description{
This function allows you to create a tag.
}
\examples{
user <- login("demo@labstep.com","demopassword")
tag <- createTag(user,'MyTag')
print(tag)
}
