% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWorkspace.R
\name{createWorkspace}
\alias{createWorkspace}
\title{Create a Workspace on Labstep}
\usage{
createWorkspace(user, name)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{name}{The name you want to give your Workspace.}
}
\value{
Returns an `workspace` object
}
\description{
This function allows you to create a Labstep Workspace.
}
\examples{
user <- login("demo@labstep.com","demopassword")
workspace <- createWorkspace(user,'My First Workspace')
print(workspace)
}
