% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiment.R
\name{getExperiment}
\alias{getExperiment}
\title{Get Experiment}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `authenticate` command}

\item{id}{ID of the experiment to retrieve}
}
\value{
Returns an `experiment` object
}
\description{
This function allows you to retrieve metadata for a single experiment
}
\examples{
user <- authenticate("demo@labstep.com","demopassword")
experiment <- getExperiment(user,123)
print(experiments)
}
