% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperiments}
\alias{getExperiments}
\title{Get List of Labstep Experiments}
\usage{
getExperiments(user, count = 100)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{count}{Number of results to return. Defaults to 100.}
}
\value{
Returns a list of `experiment` objects
}
\description{
This function allows you to search your Experiments on Labstep.
}
\examples{
user <- login("demo@labstep.com","demopassword")
experiments <- getExperiments(user)
print(experiments)
}
