% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProject.R
\name{getProject}
\alias{getProject}
\title{Retrieve metadata for a particular Labstep Project}
\usage{
getProject(user, id)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{id}{ID of the project to retrieve}
}
\value{
Returns a `project` object
}
\description{
This function allows you to retrieve metadata for a single project
}
\examples{
user <- login("demo@labstep.com","demopassword")
project <- getProject(user,1186)
print(project)
}
