% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProtocols.R
\name{getProtocols}
\alias{getProtocols}
\title{Get Protocols}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `authenticate` command}

\item{count}{Number of results to return. Defaults to 100.}
}
\value{
Returns a list of `protocol` objects
}
\description{
This function allows you to search your Protocol library on Labstep.
}
\examples{
user <- authenticate("demo@labstep.com","demopassword")
protocols <- getProtocols(user)
print(protocols)
}
