% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResource.R
\name{getResource}
\alias{getResource}
\title{Retrieve metadata for a particular Labstep Resource}
\usage{
getResource(user, id)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{id}{ID of the resource to retrieve}
}
\value{
Returns a `resource` object
}
\description{
This function allows you to retrieve metadata for a single resource
}
\examples{
user <- login("demo@labstep.com","demopassword")
resource <- getResource(user,391521)
print(resource)
}
