% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResources.R
\name{getResources}
\alias{getResources}
\title{Get Resources}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `authenticate` command}

\item{count}{Number of results to return. Defaults to 100.}
}
\value{
Returns a list of `resource` objects
}
\description{
This function allows you to search your Resource library on Labstep.
}
\examples{
user <- authenticate("demo@labstep.com","demopassword")
resources <- getResources(user)
print(resources)
}
