% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTags.R
\name{getTags}
\alias{getTags}
\title{Retrieve a list of user tags.}
\usage{
getTags(user, name = NULL, count = 100)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{name}{Optionally specify the name of tags to retrieve.}

\item{count}{Max number of results to return. Defaults to 100.}
}
\value{
Returns a list of tags with matching names.
}
\description{
This function allows you to retrieve a list of the users tags.
}
\examples{
user <- login("demo@labstep.com","demopassword")
tags <- getTags(user)
print(tags)
}
