% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWorkspace.R
\name{getWorkspace}
\alias{getWorkspace}
\title{Retrieve metadata for a particular Labstep Workspace}
\usage{
getWorkspace(user, id)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{id}{ID of the workspace to retrieve}
}
\value{
Returns a `workspace` object
}
\description{
This function allows you to retrieve metadata for a single workspace
}
\examples{
user <- login("demo@labstep.com","demopassword")
workspace <- getWorkspace(user,1186)
print(workspace)
}
