% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Authenticate a Labstep user}
\usage{
login(username, password)
}
\arguments{
\item{username}{Your Labstep username (usually the email address you signed up with)}

\item{password}{Your Labstep password}
}
\value{
Returns a 'user' object.
}
\description{
This function allows you to retrieve the user details for your Labstep account. It also returns your unique API key which you need to use the other Labstep API functions.
}
\examples{
user <- login("demo@labstep.com","demopassword")
print(user)
}
\keyword{apikey}
\keyword{login}
