% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\title{Simple tagging of a Labstep entity (creates a new tag if none exists)}
\usage{
tag(user, entity, name)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{entity}{The Labstep entity to tag. Can be a Resource, Experiment or Protocol.}

\item{name}{The name of the tag to attach.}
}
\value{
Returns the tagged entity.
}
\description{
This function allows you to tag Labstep entities. If the tag exists already it will attach it to the entity, if not a new tag will be created.
}
\examples{
user <- login("demo@labstep.com","demopassword")
experiments <- getExperiments(user,count=1)
experiment = experiments[[1]]
experiment = tag(user,experiment,'My New Tag')
}
