% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperiments}
\alias{getExperiments}
\title{Get Experiments}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `authenticate` command}
}
\value{
Returns a list of `experiment` objects
}
\description{
This function allows you to search your Protocol library on Labstep.
}
\examples{
user <- authenticate("demo@labstep.com","demopassword")
experiments <- getExperiments(user)
print(experiments)
}
