% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProtocol.R
\name{getProtocol}
\alias{getProtocol}
\title{Get Protocol}
\usage{
getProtocol(user, id)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api_key` field. Returned from `login` command}

\item{id}{ID of the protocol to retrieve}
}
\value{
Returns a `protocol` object
}
\description{
This function allows you to retrieve metadata for a single protocol
}
\examples{
user <- login("demo@labstep.com","demopassword")
protocol <- getProtocol(user,123)
print(protocol)
}
