% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResources.R
\name{getResources}
\alias{getResources}
\title{Get Resources}
\usage{
getResources(user, count = 100)
}
\arguments{
\item{user}{A labstep user object. Must contain an `api\_key` field. Returned from `login` command}

\item{count}{Number of results to return. Defaults to 100.}
}
\value{
Returns a list of `resource` objects
}
\description{
This function allows you to search your Resource library on Labstep.
}
\examples{
user <- login("demo@labstep.com","demopassword")
resources <- getResources(user)
print(resources)
}
